100-get_endianness.c

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -9,5 +9,5 @@ int get_endianness(void)
 	unsigned int n = 1;
 	char *ch = (char *)&n;
 
-	return ((int)*ch);
+	return (*ch);
 }
